ApiEmpleados/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.2.2
fastapi==0.103.1
fastapi-cli==0.0.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.8
importlib-metadata==1.5.0
itsdangerous==2.1.2
Jinja2==3.1.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.9.7
psycopg2==2.9.9
pydantic==2.5.2
pydantic-extra-types==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.14.5
Pygments==2.17.2
python-consul==1.1.0
python-decouple==3.8
python-dotenv==0.21.1
python-multipart==0.0.8
PyYAML==6.0.1
redis==5.0.8
requests==2.31.0
rich==13.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typer==0.12.5
typing_extensions==4.7.1
ujson==5.7.0
urllib3==2.0.7
uvicorn==0.22.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
